complete initial setup
