+++ .eslintrc
@@ -1,0 +1,19 @@
+{
+  "parserOptions": {
+    "ecmaVersion": 6,
+    "sourceType": "module"
+  },
+  "extends": ["airbnb-base", "react-app"],
+  "globals": {
+    "document": true,
+    "window": true,
+    "allowTemplateLiterals": true
+  },
+  "rules": {
+    "max-len": [2, 120, 4],
+    "no-console": [1, { "allow": ["error"] }],
+    "no-debugger": 1,
+    "class-methods-use-this": 0,
+    "no-plusplus": ["error", { "allowForLoopAfterthoughts": true }]
+  }
+}
