+++ .gitignore
@@ -11,6 +11,12 @@
 # production
 /build
 
+# API
+/src/helpers/apikeys.js
+
+# Firebase
+/.firebase/*
+
 # misc
 .DS_Store
 .env.local
