+++ src/App/App.scss
@@ -1,3 +1,5 @@
+@import '../styles/variables';
+
 .App {
   text-align: center;
 }
@@ -9,7 +11,7 @@
 }
 
 .App-header {
-  background-color: #282c34;
+  background-color: $backgroundColor; //#282c34;
   min-height: 100vh;
   display: flex;
   flex-direction: column;
