+++ src/helpers/apiKeys.js.example
@@ -1,0 +1,13 @@
+const apiKeys = {
+  firebaseConfig: {
+    apiKey: '',
+    authDomain: '',
+    databaseURL: '',
+    projectId: '',
+    storageBucket: '',
+    messagingSenderId: '',
+    appId: '',
+  }
+};
+
+export default apiKeys;
