+++ src/styles/_variables.scss
@@ -1,0 +1,1 @@
+$backgroundColor: cadetBlue;
